Refactor location autocomplete to reuse it, fixing the hook to work on load, and make the dashboard search take users to the search page with query populated
