+++ app/web/components/LocationAutocomplete/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./LocationAutocomplete";
+export * from "./LocationAutocomplete";
