+++ app/web/features/communities/events/EventForm.tsx
@@ -2,10 +2,10 @@ import { Checkbox, FormControlLabel, Typography } from "@material-ui/core";
 import classNames from "classnames";
 import Alert from "components/Alert";
 import ImageInput from "components/ImageInput";
+import LocationAutocomplete from "components/LocationAutocomplete";
 import MarkdownInput from "components/MarkdownInput";
 import PageTitle from "components/PageTitle";
 import TextField from "components/TextField";
-import LocationAutocomplete from "features/search/LocationAutocomplete";
 import { RpcError } from "grpc-web";
 import { useTranslation } from "i18n";
 import { COMMUNITIES, GLOBAL } from "i18n/namespaces";
