+++ app/web/features/search/FilterDialog.test.tsx
@@ -11,18 +11,17 @@ import {
   NUM_GUESTS,
   PROFILE_KEYWORDS,
 } from "features/search/constants";
-import useSearchFilters, {
-  parsedQueryToFilters,
-} from "features/search/useSearchFilters";
+import useRouteWithSearchFilters from "features/search/useRouteWithSearchFilters";
 import mockRouter from "next-router-mock";
 import { HostingStatus } from "proto/api_pb";
 import wrapper from "test/hookWrapper";
 import { server } from "test/restMock";
+import { parsedQueryToSearchFilters } from "utils/SearchFilters";
 
 import FilterDialog from "./FilterDialog";
 
 const Dialog = () => {
-  const searchFilters = useSearchFilters("");
+  const searchFilters = useRouteWithSearchFilters("");
   return (
     <>
       <FilterDialog
@@ -88,7 +87,7 @@ describe("FilterDialog", () => {
     userEvent.click(screen.getByRole("button", { name: APPLY_FILTER }));
 
     await waitFor(() => {
-      expect(parsedQueryToFilters(mockRouter.query)).toMatchObject(
+      expect(parsedQueryToSearchFilters(mockRouter.query)).toMatchObject(
         expectedFilters
       );
     });
@@ -154,7 +153,7 @@ describe("FilterDialog", () => {
       ).toBeNull();
       expect(numGuestsInput).toHaveValue(null);
 
-      expect(parsedQueryToFilters(mockRouter.query)).toMatchObject({});
+      expect(parsedQueryToSearchFilters(mockRouter.query)).toMatchObject({});
     });
   });
 
@@ -185,7 +184,7 @@ describe("FilterDialog", () => {
     await waitFor(() => {
       const errors = screen.getAllByText(MUST_HAVE_LOCATION);
       expect(errors).toHaveLength(3);
-      expect(parsedQueryToFilters(mockRouter.query)).toMatchObject({});
+      expect(parsedQueryToSearchFilters(mockRouter.query)).toMatchObject({});
     });
   });
 });
