+++ app/web/features/search/SearchBox.test.tsx
@@ -9,13 +9,12 @@ import {
   SEARCH_BY_KEYWORD,
   SEARCH_BY_LOCATION,
 } from "features/search/constants";
-import useSearchFilters, {
-  SearchFilters,
-} from "features/search/useSearchFilters";
+import useRouteWithSearchFilters from "features/search/useRouteWithSearchFilters";
 import mockRouter from "next-router-mock";
 import { useEffect } from "react";
 import wrapper from "test/hookWrapper";
 import { server } from "test/restMock";
+import SearchFilters from "utils/SearchFilters";
 
 import SearchBox from "./SearchBox";
 
@@ -24,7 +23,7 @@ const View = ({
 }: {
   setActive?: (value: SearchFilters) => void;
 }) => {
-  const filters = useSearchFilters("");
+  const filters = useRouteWithSearchFilters("");
   useEffect(() => {
     setActive?.(filters.active);
   }, [filters.active, setActive]);
