+++ app/web/features/search/constants.ts
@@ -15,7 +15,6 @@ export const LAST_MONTH = "Last month";
 export const LAST_WEEK = "Last week";
 export const LOCATION = "Near location";
 export const MAP_PAGE = "Map page";
-export const MUST_BE_MORE_SPECIFIC = "Please choose a more specific place";
 export const MUST_HAVE_LOCATION = "Specify a location to use this filter";
 export const NUM_GUESTS = "Number of guests";
 export const NO_USER_RESULTS = "No users found.";
@@ -23,11 +22,6 @@ export const PROFILE_KEYWORDS = "Profile keywords";
 export const SEARCH = "Search";
 export const SEARCH_BY_LOCATION = "By location";
 export const SEARCH_BY_KEYWORD = "By keyword";
-export const SEARCH_LOCATION_HINT =
-  "Press enter or click the icon to choose a location";
-export const SEARCH_LOCATION_BUTTON = "Search location";
-export const SELECT_LOCATION =
-  "Press enter or click the icon, then select a location from the list";
 export const SHOWING_ALL = "Showing all users";
 
 export const lastActiveOptions = [
