+++ app/web/routes.ts
@@ -4,12 +4,9 @@ The source of truth for URLs is
 Please make sure this file stays in sync with that file as well as
 //app/backend/src/couchers/urls.py
 */
-import {
-  filtersToSearchQuery,
-  SearchFilters,
-} from "features/search/useSearchFilters";
 import { ReferenceType } from "proto/references_pb";
 import { ReferenceTypeStrings } from "service/references";
+import SearchFilters, { parseSearchFiltersToQuery } from "utils/SearchFilters";
 
 export const baseRoute = "/";
 
@@ -117,7 +114,7 @@ export const friendsRoute = `${connectionsRoute}/friends`;
 
 export const searchRoute = "/search";
 export const routeToSearch = (filters: SearchFilters) =>
-  `${searchRoute}?${filtersToSearchQuery(filters)}`;
+  `${searchRoute}?${parseSearchFiltersToQuery(filters)}`;
 
 export const jailRoute = "/restricted";
 export const tosRoute = "/terms";
